add: New task type: TEXT ANALYZER - European Parliament - Cultural Heritage
The goal of this task is to explore the European Parliament Datasets on Cultural Heritage. This exploration is a categorization of the parliamentary debates. This categorization is done based on keywords within the text of the debates. So that, if a debate contains a keyword, the debate will be assigned the keyword category.

The processing of the parliamentary debates is composed of the followings phases:
- Definition of the list of keywords.
- The extraction process by means of natural language processing tools using the GATE tool. 
- The visualization of the result with the libraries D3.js and NVD3.js.